add length to the print_in , print_unsigned , print_octal, hex

+++ functions.c
@@ -185,7 +185,7 @@ int print_hex(va_list args, int uppercase, int flags , int length )
  * @n: unsigned number
  * Return: count of characters
  */
-int print_unsigned_number(unsigned long  n)
+int print_unsigned_number(unsigned int  n)
 {
     int count = 0;
 
@@ -201,7 +201,7 @@ int print_unsigned_number(unsigned long  n)
  * @n: unsigned integer
  * Return: number of characters printed
  */
-int print_octal_number(unsigned long  n)
+int print_octal_number(unsigned int  n)
 {
     int count = 0;
     char buffer[12];
@@ -231,7 +231,7 @@ int print_octal_number(unsigned long  n)
  * @uppercase: 1 for uppercase, 0 for lowercase
  * Return: number of characters printed
  */
-int print_hex_number(unsigned long  n, int uppercase)
+int print_hex_number(unsigned int  n, int uppercase)
 {
     int count = 0;
     char buffer[9];
